+++ src/actions/__mocks__/electron-store.js
