+++ src/actions/__mocks__/electron.js
