+++ src/components/CoreInfoView.tsx
