+++ src/components/CoreView.tsx
