+++ src/components/DeviceInfoView.tsx
