+++ src/components/DeviceSelector.jsx
