+++ src/components/McuUpdateDialogView.tsx
