+++ src/components/MemoryBoxView.tsx
