+++ src/components/MemoryView.tsx
