+++ src/components/RegionInfoView.tsx
