+++ src/reducers/settingsReducer.ts
