+++ src/util/__mocks__/nrfconnect/core.js
