+++ src/util/__tests__/hex/s140_nrf52840_5.0.0-2.alpha_softdevice.hex
