+++ src/util/devices.ts
