+++ src/util/initPacket.ts
