+++ src/util/promise.ts
